+++ packages/fast-cli/src/cli.spec.ts
@@ -513,48 +513,5 @@ test.describe("CLI", () => {
                 ).not.toThrow();
             });
         });
-        test.describe("progress-ring", () => {
-            test.beforeAll(() => {
-                setup();
-                execSync(`cd ${tempDir} && npm run fast:init`);
-                setup();
-                execSync(`cd ${tempDir} && npm run fast:add-foundation-component:progress-ring`);
-            });
-            test.afterAll(() => {
-                teardown();
-            });
-            test("should copy files from the template", () => {
-                let files: Array<string> = [];
-
-                function testGeneratedFiles(folderName: string) {
-                    const tempDirContents = fs.readdirSync(path.resolve(tempDir, "src/components/test-component", folderName));
-                    const tempDirContentsWithFileTypes = fs.readdirSync(path.resolve(tempDir, "src/components/test-component", folderName), {
-                        withFileTypes: true
-                    });
-
-                    for (let i = 0, contentLength = tempDirContents.length; i < contentLength; i++) {
-                        if (tempDirContentsWithFileTypes[i].isDirectory()) {
-                            testGeneratedFiles(tempDirContents[i]);
-                        } else {
-                            files.push(
-                                folderName
-                                    ? `${folderName}/${tempDirContents[i]}`
-                                    : tempDirContents[i]
-                            );
-                        }
-                    }
-                }
-                
-                testGeneratedFiles("");
-                expect(files).toEqual(expectedGeneratedComponentTemplateFiles);
-            });
-            test("should be able to run the build", () => {
-                expect(
-                    () => {
-                        execSync(`cd ${tempDir} && npm run build`);
-                    }
-                ).not.toThrow();
-            });
-        });
     });
 });
